platform/requirements.txt
# Platform API dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Docker
docker>=7.0.0

# Billing
stripe>=8.0.0

# HTTP
httpx>=0.27.0

# Utils
pydantic>=2.9.0
pydantic-settings>=2.5.0
